requirements.txt
discord.py~=1.4.1
asyncio
youtube_dl
wikipedia
praw
requests
pillow
captcha
ffmpeg 
mcuuid==0.2
statcord.py
pyfiglet
humanize
discord.py[voice]
wikipedia